split encoders into multiple files
